+++ kodando-react-router/src/main/kotlin/kodando/react/router/HashRouter.kt
@@ -1,0 +1,11 @@
+package kodando.react.router
+
+import kodando.react.*
+
+/**
+ * Created by danfma on 04/04/17.
+ */
+
+fun hashRouter(setter: PropSetter<HashRouterProps>): ReactElement? {
+    return React.createElement(Module.HashRouterClass, propsBy(setter))
+}