+++ kodando-react-router/src/main/kotlin/kodando/react/router/Router.kt
@@ -1,0 +1,11 @@
+package kodando.react.router
+
+import kodando.react.*
+
+/**
+ * Created by danfma on 04/04/17.
+ */
+
+fun router(setter: PropSetter<RouterProps>): ReactElement? {
+    return React.createElement(Module.RouterClass, propsBy(setter))
+}