+++ kodando-react-router/src/main/kotlin/kodando/react/router/Switch.kt
@@ -1,0 +1,11 @@
+package kodando.react.router
+
+import kodando.react.*
+
+/**
+ * Created by danfma on 04/04/17.
+ */
+
+fun switch(propSetter: PropSetter<SwitchProps>): ReactElement? {
+    return React.createElement(Module.SwitchClass, propsBy(propSetter))
+}