+++ kodando-react/src/main/kotlin/kodando/react/Style.kt
@@ -1,0 +1,10 @@
+package kodando.react
+
+import kotlin.js.Json
+
+/**
+ * Created by danfma on 04/04/17.
+ */
+
+external interface Style : Json
+